MERGE: -c 31553 /trunk/Sapphire -> /branches/lwidentity-5.1
Fix compiler warnings on HP-UX and Mac OS

Bug #7842

(librpc: r32002)

+++ librpc/TestRpc/Samr.c
@@ -81,7 +81,7 @@ handle_t CreateSamrBinding(handle_t *binding, const wchar16_t *host)
         if (result == 0) {
             printf("Error: %s\n", errmsg);
         } else {
-            printf("Unknown error: %08lx\n", status);
+            printf("Unknown error: %08lx\n", (unsigned long int)status);
         }
 
         SAFE_FREE(hostname);
